backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.37
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic-settings==2.7.1
apscheduler==3.10.4
reportlab==4.2.5
email-validator==2.2.0
pytest==8.3.4
httpx==0.28.1
